[POC] Use the Section VO instead of a custom entity
Feeds the Section form with the Section Value Object instead
of using a custom object.

Requirements aren't respected, but the existing @require annotations
could be replaced with the ones from Form.

A next step would be to have 2 distinct forms for Create and Update,
so that they can have their own ValueObject.