EZP-25024: User Content REST controller
Adds an override of the REST Content Controller.

This override implements new versions of createContent, updateVersion, publishVersion
and deleteContent. For those, if the affected item is a User, it executes the request
with the user API, and responds as the Content APi would have.

This allows user management in PlatformUI to use the REST content API transparently,
without implementing the whole user API.